refactor: 추천 데이터 불러오는 로직 훅으로 분리
